FetchMapResponse working correctly.
Add some sugar to RiakString

+++ src/RiakClient/Commands/CRDT/Map.cs
@@ -57,7 +57,7 @@ namespace RiakClient.Commands.CRDT
         {
         }
 
-        public class Counter : MapOf<int>
+        public class Counter : MapOf<long>
         {
         }
